work on prototype

+++ src/cli.rs
@@ -111,7 +111,7 @@ struct CliArgs {
     log_routines: bool,
     /// The index in the OPB file to treat as the lowest variable
     #[arg(long, default_value_t = 0)]
-    first_var_idx: usize,
+    first_var_idx: u32,
 }
 
 #[derive(Copy, Clone, PartialEq, Eq, ValueEnum)]
